requirements.txt
torch>=1.13.0
torchvision>=0.14.0
pytorch-lightning>=2.0.0
diffusers>=0.20.0
easydict>=1.10
einops>=0.6.0
timm>=0.9.0
numpy>=1.23.0
scipy>=1.9.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
Pillow>=9.3.0
matplotlib>=3.7.0
seaborn>=0.13.0
pandas>=2.0.0
PyYAML>=6.0
opencv-python>=4.6.0
tensorboard>=2.13.0
